Extract setup helper logic from lib index
